tests/wpt/tests/css/css-animations/parsing/animation-range-end-invalid.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/scroll-animations-1/#propdef-animation-range">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script>
test_invalid_value("animation-range-end", "infinite");
test_invalid_value("animation-range-end", "1s 2s");
test_invalid_value("animation-range-end", "1s / 2s");

test_invalid_value("animation-range-end", "peek 50%");
test_invalid_value("animation-range-end", "50% contain");
test_invalid_value("animation-range-end", "50% cover");
test_invalid_value("animation-range-end", "50% entry");
test_invalid_value("animation-range-end", "50% enter");
test_invalid_value("animation-range-end", "50% exit");
test_invalid_value("animation-range-end", "contain contain");
test_invalid_value("animation-range-end", "none");
test_invalid_value("animation-range-end", "cover 50% enter 50%");
</script>
